dom/virtualfilesystem/test/mochitest/test_virtualfilesystem_event.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=1026350
-->
<head>
<meta charset="utf-8">
<title>Test Inputport Connection Event</title>
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1187723">Test FileSystemProvider Event
</a>
<script type="application/javascript;version=1.8">

"use strict";

SimpleTest.waitForExplicitFinish();
function MockCallback() {
}

function testEvent() {
  MockCallback.prototype = {
      onSuccess: function(requestId, value, hasMore) {},
      onError: function(requestId, error) {}
  };
  return new Promise(function(resolve, reject) {
    var Ci = SpecialPowers.Ci;

    var fileSystemProvider = navigator.fileSystemProvider;
    var _fileSystemId = "dummyId";
    var _displayName = "dummyFileSystem";
    var _writable = true;
    var _openedFilesLimit = 10;
    var _requestId = 1;
    var _operationRequestId = 1;
    fileSystemProvider.addEventListener("abortrequested", function onabort(event){
      fileSystemProvider.removeEventListener("abortrequested", onabort);
      ok(true, "got event.");
      is(event.options.fileSystemId, _fileSystemId, "file system id must be the same");
      is(event.options.requestId, _requestId, "request id must be the same");
      is(event.options.operationRequestId, _operationRequestId, "operation request id must be the same");
      fileSystemProvider.unmount({fileSystemId: _fileSystemId}).then(function() {
          resolve();
        }
      );
    });
    fileSystemProvider.mount({fileSystemId: _fileSystemId,
                              displayName: _displayName,
                              writable: _writable,
                              openedFilesLimit: _openedFilesLimit}).then(
      function() {
        ok(true, "mount ok.");
        var vfsService;
        vfsService = SpecialPowers.Cc["@mozilla.org/virtualfilesystem/virtualfilesystem-service;1"].getService(Ci.nsIVirtualFileSystemService);
        ok(vfsService, "must have vfsService");
        var vfs = vfsService.getVirtualFileSystemById(_fileSystemId);
        ok(vfs, "must have virtual file system.");
        vfs.abort(_requestId, new MockCallback());
      },
      function(aError) {
        ok(false, "Fail to mount: " + aError);
        resolve();
      }
    );
  });
}


function runTest() {
  ok(navigator.fileSystemProvider, "should have navigator.fileSystemProvider");

  testEvent()
  .then(function() {
    info("test finished");
    SimpleTest.finish();
  });
}

SpecialPowers.pushPrefEnv({"set": [["device.filesystemprovider.enabled", true],
                                    //to ignore app scope check.
                                   ["dom.ignore_webidl_scope_checks", true]]}, function() {
  SpecialPowers.pushPermissions(
    [{"type": "filesystemprovider", "allow": true, "context": document}], runTest);
});

 </script>
 </pre>
 </body>
 </html>
